requirements.txt
numpy==1.22.2
pandas==1.4.1
psycopg2-binary==2.9.3
schedule==1.1.0
python-dotenv==0.19.2